requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
simplejson==3.19.3
sqlparse==0.5.2
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
psycopg2-binary~=2.9.6
redis==5.2.1
django-redis==5.4.0